requirements.txt
rpi_ws281x==5.0.0
adafruit-circuitpython-neopixel==6.3.11
flask==3.0.3
RPi.GPIO==0.7.1
pygame==2.6.0
